pyproject.toml
[tool.poetry]
name = "emotionnote"
version = "0.1.0"
description = "个人情绪记录网站"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"